Created TextManaging Library Moved PrInpArrow(), ParseNumInt() and ParseNumDouble() to TextManaging Library
