+++ EcoEnergySolution/TextManaging/TextManaging.cs
@@ -1,0 +1,56 @@
+﻿namespace TextManaging
+{
+    public static class Text
+    {
+        private const string ErrFormatException = "EL VALOR INTRODUÏT NO ÉS UN NÚMERO";
+        private const string ErrOverflowException = "EL VALOR INTRODUÏT ÉS MASSA GRAN";
+        private const string ErrException = "HA OCORREGUT UN ERROR INESPERAT";
+
+        public static void PrInpArrow()
+        {
+            const string InputArrow = "> ";
+
+            Console.Write(InputArrow);
+        }
+        public static int ParseNumInt(string num)
+        {
+            try
+            {
+                return int.Parse(num);
+            }
+            catch (FormatException)
+            {
+                Console.WriteLine(ErrFormatException);
+            }
+            catch (OverflowException)
+            {
+                Console.WriteLine(ErrOverflowException);
+            }
+            catch (Exception)
+            {
+                Console.WriteLine(ErrException);
+            }
+            return 0;
+        }
+        public static double ParseNumDouble(string num)
+        {
+            try
+            {
+                return double.Parse(num);
+            }
+            catch (FormatException)
+            {
+                Console.WriteLine(ErrFormatException);
+            }
+            catch (OverflowException)
+            {
+                Console.WriteLine(ErrOverflowException);
+            }
+            catch (Exception)
+            {
+                Console.WriteLine(ErrException);
+            }
+            return 0d;
+        }
+    }
+}
